📚 README's docs reorg - folder structure strikes back
The Empire reorganized our docs into subdirectories.
The README didn't get the memo. Now it does.
getting-started/, development/, operations/ - oh my!

Updated: Main README.md with correct doc paths
Like finding the right shelf in the Library of Alexandria,
but the library is on fire and the shelves keep moving.